requirements.txt
pandas
python-dotenv
requests
supabase
backoff
httpx
pyuule
python-dateutil 